+++ erp-logics/src/main/java/lsfusion/erp/region/by/machinery/cashregister/fiscalvmk/ReceiptInstance.java
@@ -1,18 +1,19 @@
 package lsfusion.erp.region.by.machinery.cashregister.fiscalvmk;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 public class ReceiptInstance implements Serializable {
-    public Double sumDisc;
-    public Double sumCard;
-    public Double sumCash;
-    public Double sumTotal;
+    public BigDecimal sumDisc;
+    public BigDecimal sumCard;
+    public BigDecimal sumCash;
+    public BigDecimal sumTotal;
 
     public List<ReceiptItem> receiptSaleList;
     public List<ReceiptItem> receiptReturnList;
 
-    public ReceiptInstance(Double sumDisc, Double sumCard, Double sumCash, Double sumTotal,
+    public ReceiptInstance(BigDecimal sumDisc, BigDecimal sumCard, BigDecimal sumCash, BigDecimal sumTotal,
                            List<ReceiptItem> receiptSaleList, List<ReceiptItem> receiptReturnList) {
         this.sumDisc = sumDisc;
         this.sumCard = sumCard;
